requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
numpy==2.1.2
torch==2.8.0
torchvision==0.23.0
